requirements.txt
requests>=2.28.0
pillow>=9.0.0
numpy<2.0
opencv-python-headless>=4.8.0
open3d>=0.17.0
